+++ src/idunno.Bluesky/Feed/Model/GetLikesResponse.cs
@@ -5,6 +5,7 @@ using System.Diagnostics.CodeAnalysis;
 using System.Text.Json.Serialization;
 
 using idunno.AtProto;
+using idunno.Bluesky.Feed.Likes;
 
 namespace idunno.Bluesky.Feed.Model;
 
@@ -12,7 +13,7 @@ namespace idunno.Bluesky.Feed.Model;
 internal sealed record GetLikesResponse
 {
     [JsonConstructor]
-    public GetLikesResponse(AtUri uri, Cid? cid, ICollection<LikesLike> likes, string? cursor)
+    public GetLikesResponse(AtUri uri, Cid? cid, ICollection<Like> likes, string? cursor)
     {
         Uri = uri;
         Cid = cid;
@@ -29,7 +30,7 @@ internal sealed record GetLikesResponse
 
     [JsonInclude]
     [JsonRequired]
-    public ICollection<LikesLike> Likes { get; init; }
+    public ICollection<Like> Likes { get; init; }
 
     [JsonInclude]
     public string? Cursor { get; init; }