+++ CHANGELOG.md
@@ -29,6 +29,8 @@
   Use `UploadVideo(string fileName, byte[] video, string mimeType, CancellationToken cancellationToken)` with a MIME type of `video/mp4` instead.
 * `BlueskyAgent.UnmuteModeList` spelling has been corrected to `BlueskyAgent.UnmuteModList`.
 * `ListViewerState` has been converted to a C# Record and now has all it's properties as nullable, to match the lexicon.
+* `Feed.Likes` has been renamed to `LikesCollection` to better reflect its purpose.
+* `Feed.Like` has been moved to the `Feed.Likes` namespace, as it is only a component in the `LikesCollection` collection.
 
 ## 5.0.0 - 2026-08-24